+++ Sources/Carlos/Carlos.h
@@ -1,0 +1,7 @@
+@import Foundation;
+
+//! Project version number for Carlos.
+FOUNDATION_EXPORT double CarlosVersionNumber;
+
+//! Project version string for Carlos.
+FOUNDATION_EXPORT const unsigned char CarlosVersionString[];